requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
starlette==0.27.0

sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

redis==5.0.1
hiredis==2.3.2

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
pre-commit==3.6.0

python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

transformers>=4.36.0
torch>=2.6.0
sentencepiece>=0.1.99

pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
yfinance==0.2.35
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0

prefect==2.14.21
prefect-sqlalchemy==0.4.1
# Starlette pinned for FastAPI-Prefect compatibility

prometheus-client==0.19.0
ipython==8.20.0
jupyter==1.0.0
